[Feature #166148029] Rename files the right way
